# docs/swagger.yaml
definitions:
...
  exception.AppErr:
    properties:
      message:
        example: some error
        type: string
      status:
        example: 404
        type: integer
    type: object
  exception.ValidationError:
    properties:
      errors:
        items:
          $ref: '#/definitions/exception.validationErrorItem'
        type: array
      status:
        example: 400
        type: integer
    type: object
  exception.validationErrorItem:
    properties:
      key:
        example: email
        type: string
      message:
        example: email is invalid
        type: string
    type: object
...
  option.AddOptionToProductModelDto:
    properties:
      option_id:
        minimum: 1
        type: integer
      product_model_id:
        minimum: 1
        type: integer
      value_id:
        minimum: 1
        type: integer
    required:
    - option_id
    - product_model_id
    - value_id
    type: object
  option.AddSizeToProductModelDto:
    properties:
      in_stock:
        minimum: 0
        type: integer
      product_model_id:
        minimum: 1
        type: integer
      size_id:
        minimum: 1
        type: integer
    required:
    - in_stock
    - product_model_id
    - size_id
    type: object
  option.CatalogFilters:
    properties:
      options:
        items:
          $ref: '#/definitions/option.CatalogOption'
        type: array
      sizes:
        items:
          $ref: '#/definitions/option.CatalogSize'
        type: array
    type: object
  option.CatalogOption:
    properties:
      option_id:
        type: integer
      slug:
        type: string
      title:
        type: string
      values:
        items:
          $ref: '#/definitions/option.CatalogValue'
        type: array
    type: object
  option.CatalogSize:
    properties:
      size_id:
        type: integer
      value:
        type: string
    type: object
  option.CatalogValue:
    properties:
      option_id:
        type: integer
      value:
        type: string
      value_id:
        type: integer
    type: object
  option.CreateOptionDto:
    properties:
      slug:
        example: color
        minLength: 3
        type: string
      title:
        example: Цвет
        minLength: 3
        type: string
    required:
    - slug
    - title
    type: object
  option.CreateOptionValueDto:
    properties:
      info:
        type: string
      option_id:
        example: 4
        type: integer
      value:
        example: Желтый
        type: string
    type: object
  option.CreateSizeDto:
    properties:
      literal:
        type: string
      numeric:
        type: string
    required:
    - literal
    - numeric
    type: object
  option.Option:
    properties:
      id:
        example: 4
        type: integer
      slug:
        example: color
        type: string
      title:
        example: Цвет
        type: string
      values:
        items:
          $ref: '#/definitions/option.OptionValue'
        type: array
    type: object
  option.OptionValue:
    properties:
      id:
        example: 44
        type: integer
      info:
        type: string
      option_id:
        example: 4
        type: integer
      value:
        example: Желтый
        type: string
    type: object
  option.UpdateOptionDto:
    properties:
      title:
        example: Цвет
        minLength: 3
        type: string
    required:
    - title
    type: object
...
info:
  contact: {}
paths:
...
  /api/option/:
    get:
      consumes:
      - application/json
      description: Get all options
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/option.Option'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get all options
      tags:
      - option
    post:
      consumes:
      - application/json
      description: Create option with dto
      parameters:
      - description: Create option dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/option.CreateOptionDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create option
      tags:
      - option
  /api/option/:id:
    delete:
      consumes:
      - application/json
      description: Delete option by id
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Delete option by id
      tags:
      - option
    get:
      consumes:
      - application/json
      description: Get option by id
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/option.Option'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get option by id
      tags:
      - option
    patch:
      consumes:
      - application/json
      description: Update option with dto
      parameters:
      - description: Update option dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/option.UpdateOptionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Update option
      tags:
      - option
  /api/option/add-to-product-model:
    post:
      consumes:
      - application/json
      description: Add option to product model
      parameters:
      - description: Add to product model dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/option.AddOptionToProductModelDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Add option to product model
      tags:
      - option
  /api/option/catalog-filters/:categorySlug:
    get:
      consumes:
      - application/json
      description: Get Catalog Filters
      parameters:
      - description: categorySlug parameter
        in: path
        name: categorySlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/option.CatalogFilters'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get get Catalog Filters by category slug
      tags:
      - option
  /api/option/size:
    post:
      consumes:
      - application/json
      description: Create size with dto
      parameters:
      - description: Create size dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/option.CreateSizeDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create size
      tags:
      - option
  /api/option/size/add-to-product-model:
    post:
      consumes:
      - application/json
      description: Add size to product model
      parameters:
      - description: Add size to product model dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/option.AddSizeToProductModelDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Add size to product model
      tags:
      - option
  /api/option/value:
    post:
      consumes:
      - application/json
      description: Create value with dto
      parameters:
      - description: Create option value dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/option.CreateOptionValueDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create value
      tags:
      - option
  /api/option/value:id:
    delete:
      consumes:
      - application/json
      description: Delete option value by id
      parameters:
      - description: id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Delete option value by id
      tags:
      - option
...
swagger: "2.0"